rspec途中。でも、他の人のプルリクのコンフリクト解消のためこのブランチをプルリク。
