+++ app/javascript/packs/users/articles_new_and_edit.js
@@ -46,6 +46,7 @@ $(function(){
   });
   
   $('.markdown-editor').on('drop', function(e) { //dropのイベントをハンドル
+    console.warn('a')
     e.preventDefault(); //元の動きを止める処理
     var image = e.originalEvent.dataTransfer.files[0]; //ドロップされた画像の1件目を取得
     var formData = new FormData();
