--- spec/support/login_support.rb
@@ -1,8 +1,0 @@
-module LoginSupport
-  def sign_in(user)
-    visit new_user_session_path
-    fill_in 'user[email]', with: user.email
-    fill_in 'user[password]', with: user.password
-    click_button 'ログイン'
-  end
-end
